#115 実装: 購入金額を合計金額（単価×有効点数）で表示
