+++ source/app/Http/Controllers/TicketPurchaseController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\TicketPurchase\StoreTicketPurchaseRequest;
 use App\Models\TicketPurchase;
+use App\UseCases\TicketPurchase\ExpandSelectionsAction;
 use App\UseCases\TicketPurchase\StoreAction;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -12,7 +13,7 @@ use Inertia\Response;
 
 class TicketPurchaseController extends Controller
 {
-    public function index(Request $request): Response
+    public function index(Request $request, ExpandSelectionsAction $expandSelections): Response
     {
         $paginator = TicketPurchase::query()
             ->where('ticket_purchases.user_id', $request->user()->id)
@@ -29,6 +30,7 @@ class TicketPurchaseController extends Controller
                 'races.race_date',
                 'venues.name as venue_name',
                 'races.race_number',
+                'ticket_types.name as ticket_type_name',
                 'ticket_types.label as ticket_type_label',
                 'buy_types.name as buy_type_name',
             ])
@@ -48,7 +50,13 @@ class TicketPurchaseController extends Controller
             'ticket_type_label' => $purchase->ticket_type_label,
             'buy_type_name' => $purchase->buy_type_name,
             'selections' => $purchase->selections,
-            'amount' => $purchase->amount,
+            'amount' => $purchase->amount !== null
+                ? $purchase->amount * count($expandSelections->execute(
+                    $purchase->ticket_type_name,
+                    $purchase->buy_type_name,
+                    $purchase->selections,
+                ))
+                : null,
             'payout_amount' => $purchase->payout_amount !== null ? (int) $purchase->payout_amount : null,
         ]);
